Se desarrolla metodo post

+++ RecordTimeOso.Common/Models/RecordTime.cs
@@ -8,7 +8,7 @@ namespace RecordTimeOso.Common.Models
     {
         public int IdEmployee { get; set; }
         public DateTime TimeRecorded { get; set; }
-        public Byte RecordTipe { get; set; }
+        public int RecordTipe { get; set; }
         public bool CoConsolidated { get; set; }
     }
 }
